requirements.txt
PyQt6>=6.6.0
ollama>=0.2.0
httpx>=0.27.0
pyyaml>=6.0.1
